+++ docs/changelog.rst
@@ -4,6 +4,20 @@ Changelog
 0.9
 ---
 
+* PR #184, #185 and #186 have added support for Ticker fields
+  ``rtTradeVolume``, ``auctionVolume``, ``auctionPrice`` and
+  ``auctionImbalance``.
+
+Version 0.9.56
+^^^^^^^^^^^^^^
+
+* Fix bug #178: ``Order.totalQuantity`` is now float.
+
+Version 0.9.55
+^^^^^^^^^^^^^^
+
+* Sphinx update for documentation.
+
 Version 0.9.54
 ^^^^^^^^^^^^^^